Ignoring HTTP 204
When the tile doesn't have data, it has an error. We use 204 to Show the tilelayer that it's empty. So i think you should ignore 204. I'm not sure if this is the correct way to ignore it. But it's how i fixed it in my Code. Thank you for this cool  Library